Move transaction method to order decorator
