--- app/decorators/controllers/solidus_acima/checkout_controller_decorator.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-module SolidusAcima
-  module CheckoutControllerDecorator
-    def self.prepended(base)
-      base.helper ::SolidusAcima::CheckoutHelper
-    end
-
-    ::Spree::CheckoutController.prepend(self) if SolidusSupport.frontend_available?
-    ::Spree::Admin::PaymentsController.prepend(self) if SolidusSupport.backend_available?
-  end
-end
