+++ app/views/spree/shared/_acima.html.erb
@@ -2,7 +2,7 @@
   id="iframe-container" class="acima-iframe-hidden"
   data-merchant-id="<%= payment_method.preferred_merchant_id %>"
   data-iframe-url="<%= payment_method.preferred_iframe_url %>"
-  data-transaction="<%= order_for_acima_transaction(@order) %>"
+  data-transaction="<%= @order.acima_transaction %>"
   data-order-number="<%= @order.number %>"
   data-order-token="<%= @order.guest_token %>"
   data-payment-method-id="<%= payment_method.id %>"
